Refactor: Move DefaultMotion content into separate component
