+++ src/modules/dashboard/components/ActionsPage/ActionPageMotionContent/ActionPageMotionContent.css.d.ts
@@ -1,0 +1,3 @@
+export const heading: string;
+export const annotation: string;
+export const commentBox: string;
