+++ src/modules/dashboard/components/ActionsPage/ActionPageMotionContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ActionPageMotionContent';
